dev-rel if it exists

+++ Personal/910-install-personal-settings-bookmarks.sh
@@ -50,9 +50,12 @@ echo "################### Bookmarks to install"
 echo "################################################################"
 tput sgr0
 echo
-if grep -q "arconet" /etc/dev-rel || grep -q "arcopro" /etc/dev-rel || grep -q "arcoplasma" /etc/dev-rel; then
-    [ -d $HOME"/.config/gtk-3.0" ] || mkdir -p $HOME"/.config/gtk-3.0"
-    cp $installed_dir/settings/bookmarks-arco/bookmarks* ~/.config/gtk-3.0/
+
+if [ -f /etc/dev-rel ]; then
+    if grep -q "arconet" /etc/dev-rel || grep -q "arcopro" /etc/dev-rel || grep -q "arcoplasma" /etc/dev-rel; then
+        [ -d $HOME"/.config/gtk-3.0" ] || mkdir -p $HOME"/.config/gtk-3.0"
+        cp $installed_dir/settings/bookmarks-arco/bookmarks* ~/.config/gtk-3.0/
+    fi
 fi
 
 if [ -f /usr/local/bin/get-nemesis-on-alci ]; then
